requirements.txt
onnxruntime>=1.10.0
numpy>=1.19.5
Pillow>=8.4.0
torch>=1.9.0
transformers>=4.12.0
opencv-python>=4.5.5.64
matplotlib>=3.5.0
scikit-learn>=1.0.0
pyyaml>=6.0
tqdm>=4.62.0
